refactor: refractor follow control with context
