+++ src/pages/Setting.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import SideBar from '../components/SideBar/SideBar'
 import SettingForm from '../components/settingForm/SettinForm'
-import ModalControlContextProvider from '../contexts/modalControlContext/ModalControlContext'
+import { ModalControlContextProvider } from '../contexts/modalControlContext/ModalControlContext'
 
 import style from './Setting.module.scss'
